breaking: remove the cypress/angular-signals test harness. The changes incorporated in @cypress/angular-signals are now merged upstream into @cypress/angular and the package is no longer needed [run ci]
