+++ .circleci/cache-version.txt
@@ -1,3 +1,3 @@
 # Bump this version to force CI to re-create the cache from scratch.
 
-11-6-24
+11-06-24-angular-signals-removal
