+++ cli/.eslintignore
@@ -10,7 +10,6 @@ package.json
 
 # these are all copied from dist'd builds from the individual libs
 /angular
-/angular-signals
 /react
 /react18
 /vue
