+++ cli/.gitignore
@@ -20,4 +20,3 @@ react*
 mount-utils
 angular
 svelte
-angular-signals
