+++ cli/package.json
@@ -68,7 +68,6 @@
     "@babel/cli": "7.24.8",
     "@babel/preset-env": "7.25.3",
     "@cypress/angular": "0.0.0-development",
-    "@cypress/angular-signals": "0.0.0-development",
     "@cypress/grep": "0.0.0-development",
     "@cypress/mount-utils": "0.0.0-development",
     "@cypress/react": "0.0.0-development",
@@ -117,8 +116,7 @@
     "react",
     "react18",
     "angular",
-    "svelte",
-    "angular-signals"
+    "svelte"
   ],
   "bin": {
     "cypress": "bin/cypress"
@@ -165,11 +163,6 @@
       "types": "./svelte/dist/index.d.ts",
       "import": "./svelte/dist/cypress-svelte.esm-bundler.js",
       "require": "./svelte/dist/cypress-svelte.cjs.js"
-    },
-    "./angular-signals": {
-      "types": "./angular-signals/dist/index.d.ts",
-      "import": "./angular-signals/dist/index.js",
-      "require": "./angular-signals/dist/index.js"
     }
   },
   "workspaces": {
