--- npm/angular-signals/.eslintignore
@@ -1,5 +1,0 @@
-**/dist
-**/*.d.ts
-**/package-lock.json
-**/tsconfig.json
-**/cypress/fixtures
