--- npm/angular-signals/.releaserc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  ...require('../../.releaserc'),
-}
