--- npm/angular-signals/CHANGELOG.md
@@ -1,6 +1,0 @@
-# @cypress/angular-signals-v1.0.0 (2024-07-02)
-
-
-### Features
-
-* add Angular Signals CT Harness for Angular 17.2 and up for users to be able to use Angular Signals within their component tests ([#29621](https://github.com/cypress-io/cypress/issues/29621)) ([f2554f1](https://github.com/cypress-io/cypress/commit/f2554f12d6d1f438db898fbbc10a100ebff733ce))
