--- npm/angular-signals/README.md
@@ -1,11 +1,0 @@
-# @cypress/angular-signals
-
-Mount Angular components in the open source [Cypress.io](https://www.cypress.io/) test runner. This package is an extension of `@cypress/angular`, but with [signals](https://angular.dev/guide/signals) support.
-
-> **Note:** This package is bundled with the `cypress` package and should not need to be installed separately. See the [Angular Component Testing Docs](https://docs.cypress.io/guides/component-testing/angular/overview) for mounting Angular components. Installing and importing `mount` from `@cypress/angular-signals` should only be done for advanced use-cases.
-
-## Development
-
-Run `yarn build` to compile and sync packages to the `cypress` cli package.
-
-## [Changelog](./CHANGELOG.md)
