--- npm/angular-signals/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "@cypress/angular-signals",
-  "version": "0.0.0-development",
-  "description": "Test Angular Components using Signals with Cypress",
-  "main": "dist/index.js",
-  "scripts": {
-    "prebuild": "rimraf dist",
-    "build": "rollup -c rollup.config.mjs",
-    "postbuild": "node ../../scripts/sync-exported-npm-with-cli.js",
-    "check-ts": "tsc --noEmit",
-    "dev": "rollup -c rollup.config.mjs -w",
-    "lint": "eslint --ext .js,.ts,.json, ."
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@angular/common": "^17.2.0",
-    "@angular/core": "^17.2.0",
-    "@angular/platform-browser-dynamic": "^17.2.0",
-    "@cypress/mount-utils": "0.0.0-development",
-    "typescript": "~5.4.5",
-    "zone.js": "~0.14.6"
-  },
-  "peerDependencies": {
-    "@angular/common": ">=17.2",
-    "@angular/core": ">=17.2",
-    "@angular/platform-browser-dynamic": ">=17.2",
-    "rxjs": ">=7.5.0",
-    "zone.js": ">=0.13.0"
-  },
-  "files": [
-    "dist"
-  ],
-  "types": "dist/index.d.ts",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/cypress-io/cypress.git"
-  },
-  "homepage": "https://github.com/cypress-io/cypress/blob/develop/npm/angular-signals/#readme",
-  "bugs": "https://github.com/cypress-io/cypress/issues/new?assignees=&labels=npm%3A%20%40cypress%2Fangular&template=1-bug-report.md&title=",
-  "keywords": [
-    "angular",
-    "cypress",
-    "cypress-io",
-    "test",
-    "testing"
-  ],
-  "contributors": [
-    {
-      "name": "Bill Glesias",
-      "social": "@atofstryker"
-    }
-  ],
-  "module": "dist/index.js",
-  "publishConfig": {
-    "access": "public"
-  },
-  "nx": {
-    "targets": {
-      "build": {
-        "outputs": [
-          "{workspaceRoot}/cli/angular-signals"
-        ]
-      }
-    }
-  },
-  "standard": {
-    "globals": [
-      "Cypress",
-      "cy",
-      "expect"
-    ]
-  }
-}
