--- npm/angular-signals/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "experimentalDecorators": true,
-    "target": "es2020",
-    "module": "es2020",
-    "skipLibCheck": true,
-    "lib": [
-      "ESNext",
-      "DOM"
-    ],
-    "allowJs": true,
-    "declaration": true,
-    "outDir": "dist",
-    "strict": true,
-    "baseUrl": "./",
-    "types": [
-      "cypress"
-    ],
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "noPropertyAccessFromIndexSignature": true,
-  },
-  "include": ["src/**/*.*"],
-  "exclude": ["src/**/*-spec.*"]
-}
